refactor: extract assertion helpers and split analyser spec (H1)
Extract expectMetricResult, checkMetricInvariant, and FLOAT_TOLERANCE to
src/test/dataAnalysis/averagingAnalyserAssertions.ts. Split the 1959-line
averagingAnalyser.spec.ts into 4 co-located spec files (filters,
accumulation, rows, orchestration), all under 550 lines. All 31 tests
preserved verbatim.
